fix(tests): silence MSVC C4267 size_t narrowings (#23 Phase 2b)
Eight call sites across three test files where size_t flowed into a
narrower destination, triggering MSVC C4267. Two patterns:

(1) test_memory_controller.cpp:140,147 and test_dma_engine_process.cpp:365,370
    store count_events()'s size_t return into a local 'int'. Widened the
    locals to size_t so the type matches the source. No casting needed,
    and the subsequent REQUIRE(x == 1) still compiles cleanly.

(2) test_dma_memory_noc_integration.cpp:460,498,537,551 compute a node
    id from a size_t loop variable and assign it to DMATransfer::dst_id,
    which is uint32_t (dma_engine_interface.hpp:60). The truncation is
    safe in practice (these tests use 4x4 = 16-node grids; the value
    never approaches 2^32) but should be made explicit with
    static_cast<uint32_t> at the assignment.

Verified locally on Linux gcc release:
  test_memory_controller:               12/12 cases,  51/51 assertions
  test_dma_engine_process:              13/13 cases,  54/54 assertions
  dma_memory_noc_integration_test:       8/8 cases,  61/61 assertions

Refs #23

+++ tests/timing/test_dma_engine_process.cpp
@@ -362,12 +362,12 @@ TEST_CASE("DMAEngineProcess sees MC command bus serialization", "[timing][dma_pr
 
     // First MC tick - only ONE command issued
     auto mc_events = mc.tick(0);
-    int load_starts_at_0 = count_events(mc_events, EventType::DMA_LOAD_START);
+    size_t load_starts_at_0 = count_events(mc_events, EventType::DMA_LOAD_START);
     REQUIRE(load_starts_at_0 == 1);
 
     // Second MC tick - other command issued
     mc_events = mc.tick(1);
-    int load_starts_at_1 = count_events(mc_events, EventType::DMA_LOAD_START);
+    size_t load_starts_at_1 = count_events(mc_events, EventType::DMA_LOAD_START);
     REQUIRE(load_starts_at_1 == 1);
 
     // Total: 2 commands serialized over 2 cycles
